LIME-42: * Icons moved
