+++ frontend/src/bundles/common/components/icon/enums/icon-components.enum.ts
@@ -1,11 +1,7 @@
 import ArrowDownIcon from '~/assets/img/icons/arrow-down-icon.svg?react';
-import ScheduleIcon from '~/assets/img/icons/calendar-days-icon.svg?react';
 import GoalsIcon from '~/assets/img/icons/goals-icon.svg?react';
-import HelpIcon from '~/assets/img/icons/help-icon.svg?react';
 import LogoIcon from '~/assets/img/icons/logo-icon.svg?react';
-import LogoutIcon from '~/assets/img/icons/logout-icon.svg?react';
 import NotFoundIcon from '~/assets/img/icons/not-found-icon.svg?react';
-import OverviewIcon from '~/assets/img/icons/overview-icon.svg?react';
 import WorkoutIcon from '~/assets/img/icons/workout-icon.svg?react';
 
 import { type IconName } from '../types/icon.type.js';
@@ -17,12 +13,9 @@ const IconComponent: Record<
     arrowDown: ArrowDownIcon,
     logoIcon: LogoIcon,
     notFoundIcon: NotFoundIcon,
-    goalsIcon: GoalsIcon,
-    overviewIcon: OverviewIcon,
-    helpIcon: HelpIcon,
-    logoutIcon: LogoutIcon,
+    goalsIcon: GoalsIcon,        
     workoutIcon: WorkoutIcon,
-    scheduleIcon: ScheduleIcon,
+    
 } as const;
 
 export { IconComponent };
