+++ frontend/src/bundles/common/components/icon/types/icon.type.ts
@@ -2,11 +2,7 @@ type IconName =
     | 'arrowDown'
     | 'logoIcon'
     | 'notFoundIcon'
-    | 'goalsIcon'
-    | 'overviewIcon'
-    | 'workoutIcon'
-    | 'helpIcon'
-    | 'logoutIcon'
-    | 'scheduleIcon';
+    | 'goalsIcon'    
+    | 'workoutIcon';
 
 export { type IconName };
